pyproject.toml
[project]
name = "academicdb"
version = "0.2.1"
description = "Database for an academic CV"
authors = [
  {name = "Russ Poldrack", email = "poldrack@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.3",
    "numpy>=2.2.1",
    "biopython>=1.84",
    "scholarly>=1.7.11",
    "crossrefapi>=1.6.0",
    "tomli>=2.2.1",
    "pybliometrics>=4.1",
    "pymongo>=4.10.1",
    "orcid>=1.0.3",
    "tomli-w>=1.1.0",
    "pytest>=8.3.4",
    "toml>=0.10.2",
    "ipython>=8.31.0",
    "matplotlib>=3.10.0",
    # Django web interface dependencies
    "django>=4.2,<5.0",
    "psycopg2-binary>=2.9.0",
    "django-allauth>=0.57.0",
    "django-extensions>=3.2.0",
    "djangorestframework>=3.14.0",
    "python-dotenv>=1.1.0",
]

[project.scripts]
dbbuilder = "academicdb.dbbuilder:main"
get_collaborators = "academicdb.get_collaborators:main"
render_cv = "academicdb.render_cv:main"
academicdb-web = "academicdb.webinterface:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "pytest-django>=4.5.0",
]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  migrations
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
skip_glob = ["*/migrations/*"]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "migrations",
    ".venv",
    "venv",
]
